feat(control): add form for lic papier
